DjangoBackend/requirements.txt
asgiref==2.3.0
Django==2.0.3
channels==2.1.1
djangorestframework==3.8.2
